requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
requests>=2.31.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
